fix(Table): fix default horizontal
