+++ src/table/hooks/useRowHighlight.ts
@@ -22,7 +22,6 @@ import {
 export function useRowHighlight(props: BaseTableProps, tableRef: Ref<HTMLDivElement>) {
   const { data, activeRowType, activeRowKeys, defaultActiveRowKeys, disableSpaceInactiveRow } = toRefs(props);
   const currentOperationRowIndex = ref(-1);
-  const tableRefTabIndex = ref(0);
   const isShiftPressed = ref(false);
   const shiftSelectionState = ref(false);
   const areaSelectionStartIndex = ref(-1);
@@ -218,13 +217,6 @@ export function useRowHighlight(props: BaseTableProps, tableRef: Ref<HTMLDivElem
     if (!SPACE_REG.test(code)) {
       clearShiftAreaSelection();
     }
-
-    // 用于支持键盘默认的左右滚动，左右滚动时重置为undefined，其他情况下为0，支持键盘操作
-    if (ARROW_LEFT_REG.test(code) || ARROW_RIGHT_REG.test(code)) {
-      tableRefTabIndex.value = undefined;
-    } else {
-      tableRefTabIndex.value = 0;
-    }
   };
 
   const keyboardUpListener = (e: KeyboardEvent) => {
@@ -249,7 +241,6 @@ export function useRowHighlight(props: BaseTableProps, tableRef: Ref<HTMLDivElem
     onHighlightRow,
     addHighlightKeyboardListener,
     removeHighlightKeyboardListener,
-    tableRefTabIndex,
   };
 }
